backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
openai==1.59.9
python-dotenv==1.0.1
